+++ src/library/DIPS.Mobile.UI/Components/Chips/Chip.cs
@@ -1,12 +1,11 @@
-using DIPS.Mobile.UI.Extensions;
-using DIPS.Mobile.UI.Resources.Colors;
 using DIPS.Mobile.UI.Resources.Styles.Chip;
-using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
 namespace DIPS.Mobile.UI.Components.Chips;
 
 public partial class Chip : View
 {
+    private static Style? m_buttonToggleStyle;
+
     public Chip()
     {
         Style = InputStyle.Current;
@@ -22,5 +21,22 @@ public partial class Chip : View
     {
         CloseCommand?.Execute(CloseCommandParameter);
         CloseTapped?.Invoke(this, EventArgs.Empty);
+    }
+
+    private void OnIsToggledChanged()
+    {
+        if (!IsToggled)
+        {
+            if (m_buttonToggleStyle is not null)
+            {
+                Style = m_buttonToggleStyle;
+            }
+        }
+        else
+        {
+            m_buttonToggleStyle ??= Style;
+
+            Style = ToggleStyle.ToggledOn;
+        }
     }
 }