+++ src/library/DIPS.Mobile.UI/Components/Chips/ChipHandler.cs
@@ -1,8 +1,12 @@
+using DIPS.Mobile.UI.Resources.Styles;
+using DIPS.Mobile.UI.Resources.Styles.Chip;
+
 namespace DIPS.Mobile.UI.Components.Chips;
 
 public partial class ChipHandler
 {
     private IconName CloseIconName => IconName.close_line;
+    private IconName ToggledIconName => IconName.check_line;
     public ChipHandler() : base(PropertyMapper)
     {
     }
@@ -10,12 +14,15 @@ public partial class ChipHandler
     public static readonly IPropertyMapper<Chip, ChipHandler> PropertyMapper = new PropertyMapper<Chip, ChipHandler>(ViewMapper)
     {
         [nameof(Chip.Title)] = MapTitle,
-        [nameof(Chip.HasCloseButton)] = MapHasCloseButton,
+        [nameof(Chip.IsCloseable)] = MapIsCloseable,
         [nameof(Chip.Color)] = MapColor,
         [nameof(Chip.CloseButtonColor)] = MapCloseButtonColor,
         [nameof(Chip.CornerRadius)] = MapCornerRadius,
         [nameof(Chip.BorderWidth)] = MapBorderWidth,
         [nameof(Chip.BorderColor)] = MapBorderColor,
+        [nameof(Chip.IsToggled)] = MapIsToggled,
+        [nameof(Chip.TitleColor)] = MapTitleColor,
+        [nameof(Chip.IsToggleable)] = MapIsToggleable,
     };
 
     private static partial void MapBorderColor(ChipHandler handler, Chip chip);
@@ -27,8 +34,14 @@ public partial class ChipHandler
     private static partial void MapColor(ChipHandler handler, Chip chip);
 
     private static partial void MapTitle(ChipHandler handler, Chip chip);
-    private static partial void MapHasCloseButton(ChipHandler handler, Chip chip);
+    private static partial void MapIsCloseable(ChipHandler handler, Chip chip);
     private static partial void MapCloseButtonColor(ChipHandler handler, Chip chip);
+    private static partial void MapIsToggled(ChipHandler handler, Chip chip);
+    private static partial void MapTitleColor(ChipHandler handler, Chip chip);
+    private static partial void MapIsToggleable(ChipHandler handler, Chip chip);
+
+    
+    
     
     internal void OnChipTapped()
     {