+++ src/library/DIPS.Mobile.UI/Components/Chips/dotnet/ChipHandler.cs
@@ -7,7 +7,7 @@ public partial class ChipHandler : ViewHandler<Chip, ViewHandler>
 {
     private static partial void MapTitle(ChipHandler handler, Chip chip) => throw new Only_Here_For_UnitTests();
 
-    private static partial void MapHasCloseButton(ChipHandler handler, Chip chip) =>
+    private static partial void MapIsCloseable(ChipHandler handler, Chip chip) =>
         throw new Only_Here_For_UnitTests();
 
     protected override ViewHandler CreatePlatformView() => throw new Only_Here_For_UnitTests();
@@ -35,4 +35,20 @@ public partial class ChipHandler : ViewHandler<Chip, ViewHandler>
     {
         throw new Only_Here_For_UnitTests();
     }
+    
+    private static partial void MapTitleColor(ChipHandler handler, Chip chip)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
+    
+    private static partial void MapIsToggled(ChipHandler handler, Chip chip)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
+
+    private static partial void MapIsToggleable(ChipHandler handler, Chip chip)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
+
 }