+++ src/library/DIPS.Mobile.UI/Resources/Styles/Chip/ChipStyleResources.cs
@@ -5,6 +5,6 @@ internal static class ChipStyleResources
     public static Dictionary<ChipStyle, Style> Styles { get; } = new()
     {
         [ChipStyle.Input] = InputStyle.Current, 
-        [ChipStyle.EmptyInput] = EmptyInputStyle.Current
+        [ChipStyle.EmptyInput] = EmptyInputStyle.Current,
     };
 }