+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [19.0.0]
+- Added toggleable property on `Chip`
+- [BreakingChange] Change property `Chip.HasCloseButton` to `Chip.IsCloseable`
+
 ## [18.2.0]
 - Added new control: 'Entry'